added comments

+++ backend/flaskr/__init__.py
@@ -8,6 +8,7 @@ from models import setup_db, Question, Category
 
 QUESTIONS_PER_PAGE = 10
 
+# a function to help in paginating the questions
 def paginate_questions(request, selection):
     page = request.args.get('page', 1, type=int)
     start = (page - 1) * QUESTIONS_PER_PAGE
@@ -22,12 +23,12 @@ def create_app(test_config=None):
     setup_db(app)
 
     """
-    @TODO: Set up CORS. Allow '*' for origins. Delete the sample route after completing the TODOs
+    Set up CORS. Allow '*' for origins. Delete the sample route after completing the TODOs
     """
     CORS(app, resources={r"/": {"origins": "*"}})
 
     """
-    @TODO: Use the after_request decorator to set Access-Control-Allow
+    Use the after_request decorator to set Access-Control-Allow
     """
     @app.after_request
     def after_request(response):
@@ -40,9 +41,7 @@ def create_app(test_config=None):
         return response
     
     """
-    @TODO:
-    Create an endpoint to handle GET requests
-    for all available categories.
+    endpoint to handle GET requests for all available categories.
     """
     @app.route('/categories')
     def get_categories():
@@ -55,55 +54,63 @@ def create_app(test_config=None):
 
 
     """
-    @TODO:
-    Create an endpoint to handle GET requests for questions,
+    endpoint to handle GET requests for questions,
     including pagination (every 10 questions).
     This endpoint should return a list of questions,
     number of total questions, current category, categories.
-
-    TEST: At this point, when you start the application
-    you should see questions and categories generated,
-    ten questions per page and pagination at the bottom of the screen for three pages.
-    Clicking on the page numbers should update the questions.
     """
     @app.route('/questions')
     def get_questions():
-        selection = Question.query.order_by(Question.id).all()
-        current_questions = paginate_questions(request, selection)
-        if (len(current_questions) == 0):
-            abort(404)
+        try:
+            # select all questions to paginate
+            selection = Question.query.order_by(Question.id).all() 
+            current_questions = paginate_questions(request, selection)
 
-        current_category = request.args.get('category')
+            #if there are no more questions return 404
+            if (len(current_questions) == 0):
+                abort(404)
 
-        c_selection = Category.query.order_by(Category.id).all()
-        categories = [category.type for category in c_selection]
+            # get current category from args if exists
+            current_category = request.args.get('category')
 
-        return jsonify({
-            'success': True,
-            'questions': current_questions,
-            'total_questions':len(selection),
-            'categories': categories,
-            'current_category': current_category
-            })
+            # select all categories and extract the type
+            c_selection = Category.query.order_by(Category.id).all()
+            categories = [category.type for category in c_selection]
+
+            return jsonify({
+                'success': True,
+                'questions': current_questions,
+                'total_questions':len(selection),
+                'categories': categories,
+                'current_category': current_category
+                })
+        except:
+            abort(400)
     
     """
-    @TODO:
-    Create an endpoint to DELETE question using a question ID.
-
-    TEST: When you click the trash icon next to a question, the question will be removed.
-    This removal will persist in the database and when you refresh the page.
+    endpoint to DELETE question using a question ID.
     """
     @app.route('/questions/<int:question_id>', methods=['DELETE'])
     def delete_question(question_id):
         try:
             question = Question.query.filter_by(id=question_id).one_or_none()
+            # if the question is not found return 404
+            if (question is None):
+                abort(404)
 
             question.delete()
+
+            # get current category from args if exists
             current_category = request.args.get('category')
+
+            # select all questions to paginate
             selection = Question.query.order_by(Question.id).all()
             current_questions = paginate_questions(request, selection)
+
+            # select all categories and extract the type
             c_selection = Category.query.order_by(Category.id).all()
             categories = [category.type for category in c_selection]
+
             return jsonify({
             'success': True,
             'questions': current_questions,
@@ -111,34 +118,35 @@ def create_app(test_config=None):
             'categories': categories,
             'current_category': current_category
             })
+        
         except:
-            abort(404)
+            abort(400)
 
 
     """
-    @TODO:
-    Create an endpoint to POST a new question,
+    endpoint to POST a new question,
     which will require the question and answer text,
     category, and difficulty score.
-
-    TEST: When you submit a question on the "Add" tab,
-    the form will clear and the question will appear at the end of the last page
-    of the questions list in the "List" tab.
     """
     @app.route('/questions', methods=['POST'])
     def create_question():
-        body = request.get_json()
-        question = body.get('question', None)
-        answer = body.get('answer', None)
-        difficulty = body.get('difficulty', None)
-        category = body.get('category', None)
         try:
+            # get questions parameters
+            body = request.get_json()
+            question = body.get('question', None)
+            answer = body.get('answer', None)
+            difficulty = body.get('difficulty', None)
+            category = body.get('category', None)
+
+            # create new question and add to db
             question = Question(question=question, answer=answer, difficulty=difficulty, category=category)
             question.insert()
 
+            # select all questions to paginate
             questions = Question.query.order_by(Question.id).all()
             current_questions = paginate_questions(request, questions)
 
+            # select all categories and extract the type
             c_selection = Category.query.order_by(Category.id).all()
             categories = [category.type for category in c_selection]
 
@@ -151,29 +159,27 @@ def create_app(test_config=None):
                 'current_category': question.category
             })
         except:
-            abort(404)
+            abort(405)
 
     """
-    @TODO:
-    Create a POST endpoint to get questions based on a search term.
+    POST endpoint to get questions based on a search term.
     It should return any questions for whom the search term
     is a substring of the question.
-
-    TEST: Search by any phrase. The questions list will update to include
-    only question that include that string within their question.
-    Try using the word "title" to start.
     """
     @app.route('/questions/search', methods=['POST'])
     def search_questions():
-        body = request.get_json()
-        search_term = body.get("searchTerm")
-        print(search_term)
         try:
+            # get search term
+            body = request.get_json()
+            search_term = body.get("searchTerm")
+
+            # select all questions that includes the search term and paginate them
             selection = Question.query.order_by(Question.id).filter(
                         Question.question.ilike("%{}%".format(search_term))
             )
             current_questions = paginate_questions(request, selection)
 
+            # get current category from args if exists
             current_category = request.args.get('category')
 
             return jsonify(
@@ -188,17 +194,14 @@ def create_app(test_config=None):
             abort(404)
 
     """
-    @TODO:
-    Create a GET endpoint to get questions based on category.
-
-    TEST: In the "List" tab / main screen, clicking on one of the
-    categories in the left column will cause only questions of that
-    category to be shown.
+    GET endpoint to get questions based on category.
     """
     @app.route('/categories/<int:category_id>/questions')
     def get_questions_by_category(category_id):
+        # get the specific category
         category = Category.query.filter_by(id=category_id).one_or_none()
         try:
+            # select all questions that belongs to the category and paginate
             questions = Question.query.filter_by(category=str(category.id)).all()
             current_questions = paginate_questions(request, questions)
             
@@ -213,30 +216,28 @@ def create_app(test_config=None):
             abort(404)
 
     """
-    @TODO:
-    Create a POST endpoint to get questions to play the quiz.
+    POST endpoint to get questions to play the quiz.
     This endpoint should take category and previous question parameters
     and return a random questions within the given category,
     if provided, and that is not one of the previous questions.
-
-    TEST: In the "Play" tab, after a user selects "All" or a category,
-    one question at a time is displayed, the user is allowed to answer
-    and shown whether they were correct or not.
     """
     @app.route('/quizzes', methods=['POST'])
     def play_quiz():
-        body = request.get_json()
-        print(body)
-        quiz_category = body.get('quiz_category')
-        previous_questions = body.get('previous_questions', None)
-        category_id = int(quiz_category['id'])+1
         try:
+            # get all quiz info and add 1 to category id
+            body = request.get_json()
+            quiz_category = body.get('quiz_category')
+            previous_questions = body.get('previous_questions', None)
+            category_id = int(quiz_category['id'])+1
+
+            # if user selects all
             if( quiz_category['type'] == 'click'):
                 questions = Question.query.filter(Question.id.notin_(previous_questions)).all()
             
             else:
                 questions = Question.query.filter_by(category=str(category_id)).filter(Question.id.notin_(previous_questions)).all()
-                
+
+            # select a random question from the questions list    
             index = random.randint(0, len(questions)-1)
             random_question = questions[index]
             print(random_question.format())
@@ -249,9 +250,7 @@ def create_app(test_config=None):
 
 
     """
-    @TODO:
-    Create error handlers for all expected errors
-    including 404 and 422.
+    error handlers
     """
     @app.errorhandler(404)
     def not_found(error):
